Moving variable sync from fuse_publishers to fuse_variables
